(#160338474): fix validation error
validate input when orders is placed

  - modify order data structure
  - create customer validation for order
  - validate order request parameter
  - modify routes
  - rename test files

[#160338474]
