+++ Server/src/middleware/validation/validationFields.js
@@ -1,0 +1,2 @@
+export const userOrderfields = ['foodItemId', 'quantity'];
+export const userinfofields = ['userId', 'destinationAddress'];
