+++ package.json
@@ -16,7 +16,7 @@
     "start:webpack-devServer": "webpack-dev-server --open",
     "local:test": "nodemon --watch . --exec cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text npm run mocha:test",
     "test": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text npm run mocha:test",
-    "mocha:test": "mocha --compilers js:@babel/register ./Server/spec/user.test.js ./Server/spec/fastFood.test.js ./Server/spec/order.test.js --exit",
+    "mocha:test": "mocha --compilers js:@babel/register ./Server/spec/userTest.js ./Server/spec/fastFoodTest.js ./Server/spec/orderTest.js --exit",
     "start": "babel-node ./Server/server.js",
     "build:ui-template": "webpack",
     "coveralls": "nyc npm test && nyc report --reporter=text-lcov | coveralls"
@@ -43,7 +43,7 @@
     "coveralls": "^3.0.2",
     "cross-env": "^5.2.0",
     "css-loader": "^1.0.0",
-    "eslint": "^5.5.0",
+    "eslint": "^5.6.1",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
     "file-loader": "^2.0.0",
